More explicit casting (164 to go)

+++ upnp/src/gena/gena_device.c
@@ -449,7 +449,7 @@ static char *AllocGenaHeaders(
 		"%s%s%" PRIzu "%s%s%s",
 		HEADER_LINE_1,
 		HEADER_LINE_2A,
-		strlen(propertySet) + 2,
+		(unsigned long)strlen(propertySet) + 2,
 		HEADER_LINE_2B,
 		HEADER_LINE_3,
 		HEADER_LINE_4);
